test: implement POST /books tests

+++ src/domain/book/use-cases/remove-book.ts
@@ -1,5 +1,6 @@
 import { EntityNotFoundError } from "../../errors/EntityNotFoundError";
 import { ForbiddenOperationError } from "../../errors/ForbiddenOperationError";
+import { BookStatus } from "../Book";
 import { BookRepository } from "../repositories/BookRepository";
 
 
@@ -15,13 +16,14 @@ export class RemoveUseCase {
 
     async execute(input: RemoveBookUseCaseInput) {
         const book = await this.bookRepository.findBookById(input.id);
-
         if (!book) {
             throw new EntityNotFoundError('Book', input.id.toString())
         }
-
         if (input.ownerId !== book.ownerId) {
             throw new ForbiddenOperationError('Only owners can delete their books');
+        }
+        if (book.status !== BookStatus.PUBLISHED) {
+            throw new ForbiddenOperationError('only unsold books can be deleted');
         }
 
         await this.bookRepository.removeBook(input.id);
